+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.1-beta.4
+
+- Adding onVenueEnterCallback
+- Adding first implementation of Place and Venue objects
+- Adding setFloor method
+
+## 0.0.1-beta.3
+
+- Fixing README.md format issue
+
 ## 0.0.1-beta.2
 
 - Updating README.md